requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
binance-futures-connector==4.0.0
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cheroot==10.0.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
Cython==3.0.3
dateparser==1.1.8
dnspython==2.4.2
Flask==3.0.0
Flask-RESTful==0.3.10
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
jaraco.functools==3.9.0
Jinja2==3.1.2
lucit-licensing-python==1.1.9
magic-filter==1.0.12
MarkupSafe==2.1.3
more-itertools==10.1.0
motor==3.3.2
multidict==6.0.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.6.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
service-identity==23.1.0
simplejson==3.19.2
six==1.16.0
typing_extensions==4.7.1
tzlocal==5.1
ujson==5.8.0
unicorn-binance-rest-api==1.10.0
unicorn-binance-websocket-api==2.0.0
unicorn-fy==0.12.2
urllib3==2.0.6
websocket-client==1.6.4
websockets==10.4
Werkzeug==3.0.0
yarl==1.9.2
